Refactor reservationService
After this commit granularity and concurrent flights limits are fetched
from db. This commit also moves allowReservation function from
reservationService to routes/reservations.ts, where route handlers are
currently defined.
