+++ backend/src/routes/reservations.ts
@@ -1,6 +1,22 @@
 import express from 'express';
 import { createReservationValidator, updateReservationValidator, getTimeRangeValidator } from '@lentovaraukset/shared/src/validation/validation';
+import countMostConcurrent from '@lentovaraukset/shared/src/overlap';
 import reservationService from '../services/reservationService';
+import airfieldService from '../services/airfieldService';
+
+const allowReservation = async (
+  startTime: Date,
+  endTime: Date,
+  id: number | undefined,
+  maxConcurrentReservations: number,
+): Promise<boolean> => {
+  const reservations = (await reservationService.getReservationFromRange(startTime, endTime))
+    .filter((e) => e.id !== id);
+
+  const mostConcurrentReservations = countMostConcurrent(reservations);
+
+  return mostConcurrentReservations < maxConcurrentReservations;
+};
 
 const router = express.Router();
 
@@ -29,7 +45,19 @@ router.delete('/:id', async (req: express.Request, res: express.Response) => {
 
 router.post('/', async (req: express.Request, res: express.Response) => {
   try {
-    const newReservation = createReservationValidator(10).parse(req.body);
+    const airfield = await airfieldService.getAirfield(1); // TODO: get airfieldId from request
+    const newReservation = createReservationValidator(airfield.eventGranularityMinutes)
+      .parse(req.body);
+
+    if (!await allowReservation(
+      newReservation.start,
+      newReservation.end,
+      undefined,
+      airfield.maxConcurrentFlights,
+    )) {
+      throw new Error('Too many concurrent reservations');
+    }
+
     const reservation = await reservationService.createReservation(newReservation);
     res.json(reservation);
   } catch (error: unknown) {
@@ -42,7 +70,19 @@ router.post('/', async (req: express.Request, res: express.Response) => {
 router.patch('/:id', async (req: express.Request, res: express.Response) => {
   try {
     const id = Number(req.params.id);
-    const validReservationUpdate = updateReservationValidator(10).parse(req.body);
+    const airfield = await airfieldService.getAirfield(1); // TODO: get airfieldId from request
+    const validReservationUpdate = updateReservationValidator(airfield.eventGranularityMinutes)
+      .parse(req.body);
+
+    if (!await allowReservation(
+      validReservationUpdate.start,
+      validReservationUpdate.end,
+      undefined,
+      airfield.maxConcurrentFlights,
+    )) {
+      throw new Error('Too many concurrent reservations');
+    }
+
     const modifiedReservation = await reservationService.updateById(id, validReservationUpdate);
     res.status(200).json(modifiedReservation);
   } catch (error: unknown) {
